+++ bukkit/src/main/java/de/maxhenkel/voicechat/command/VoiceChatCommands.java
@@ -14,7 +14,6 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabCompleter;
 import org.bukkit.entity.Player;
-import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nullable;
 import java.util.Arrays;
@@ -59,7 +58,7 @@ public class VoiceChatCommands implements CommandExecutor, TabCompleter {
 
     @Nullable
     @Override
-    public List<String> onTabComplete(@NotNull CommandSender commandSender, @NotNull Command command, @NotNull String label, @NotNull String[] args) {
+    public List<String> onTabComplete(CommandSender commandSender, Command command, String label, String[] args) {
         if (args.length == 1) {
             return tabCompleteList(args[0], Arrays.asList("help", "test", "invite", "join", "leave"));
         }